% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrcalc.R
\name{correlation}
\alias{correlation}
\alias{corrcalc}
\title{correlation}
\usage{
corrcalc(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plots and data frames
}
\description{
correlation analysis
}
\details{
this program takes input as data frame and performs correlation analysis on the data.
}
\author{
Anushka Jain
}
