% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcacalc.R
\name{pca}
\alias{pca}
\alias{pcacalc}
\title{pca}
\usage{
pcacalc(df)
}
\arguments{
\item{df}{data frame}
}
\value{
plots and data frames
}
\description{
pc analysis
}
\details{
this program takes input as data frame and performs principal component analysis on the data.
}
\author{
Anushka Jain
}
